requirements.txt
random-words